improve performance

+++ controllers/user/fucntions.go
@@ -31,9 +31,33 @@ import (
 	"golang.org/x/crypto/bcrypt"
 )
 
+// execLogin is a function that handles the login process for a user.
+// It takes in the following parameters:
+// - ctx: The gin.Context object for handling HTTP requests and responses.
+// - username: The username of the user.
+// - password: The password of the user.
+// - clienttoken: The client token for the user.
+// - ClientID: The client ID for the user.
+// - Renew: A boolean value indicating whether the session should be renewed.
+//
+// The function performs the following steps:
+// 1. Logs the start time of the function execution.
+// 2. Defer a function to log the performance duration of the function.
+// 3. If Renew is true, it checks if the session exists in the session cache.
+//    - If the session exists, it checks if the session has expired.
+//      - If the session has expired, it renews the session and returns the updated user information.
+//      - If the session has not expired, it returns an error indicating that the session has already expired.
+//    - If the session does not exist, it returns an error indicating that the session renewal failed.
+// 4. If Renew is false, it performs the login process.
+//    - It queries the database to retrieve the user information.
+//    - If the user is found and the password matches, it updates the last sign-on date in the database.
+//    - It generates an authentication token for the user and stores it in the session cache.
+//    - It returns the user information and the authentication token.
+// 5. If any error occurs during the execution, it returns an error response.
+
 func execLogin(ctx *gin.Context, username string, password string, clienttoken string, ClientID string, Renew bool) {
 
-	log := logger.Log{ModuleName: logger.API, User: username, ClientID: ClientID, ControllerName: "UserController"}
+	log := logger.Log{ModuleName: logger.API, User: username, ClientID: ClientID, ControllerName: "UserController.execLogin"}
 
 	startTime := time.Now()
 	defer func() {
@@ -48,7 +72,7 @@ func execLogin(ctx *gin.Context, username string, password string, clienttoken s
 		}
 	}()
 
-	log.Debug("Login execution function is called.")
+	//	log.Debug("Login execution function is called.")
 	log.Debug(fmt.Sprintf("login parameters:%s  %s  token: %s  renew:%s", username, password, clienttoken, Renew))
 	fmt.Println("Session Timeout:", config.SessionCacheTimeout)
 	if Renew {
@@ -199,7 +223,7 @@ func execLogin(ctx *gin.Context, username string, password string, clienttoken s
 		log.Debug(fmt.Sprintf("ID:%d  Name:%s  FamilyName:%s", ID, Name, FamilyName))
 
 		Columns := []string{"LastSignOnDate", "UpdatedOn", "UpdatedBy"}
-		Values := []string{time.Now().Format("2006-01-02 15:04:05"), time.Now().Format("2006-01-02 15:04:05"), username}
+		Values := []string{time.Now().UTC().Format("2006-01-02 15:04:05"), time.Now().UTC().Format("2006-01-02 15:04:05"), username}
 		datatypes := []int{0, 0, 0}
 		Wherestr := fmt.Sprintf("ID= %d", ID)
 
@@ -241,21 +265,25 @@ func execLogin(ctx *gin.Context, username string, password string, clienttoken s
 
 }
 
+// getUserImage retrieves the user's image URL from the database.
+// It takes the username and client ID as parameters and returns the image URL as a string.
+// If an error occurs during the execution, it returns an error.
+
 func getUserImage(username string, clientid string) (string, error) {
-	log := logger.Log{ModuleName: logger.API, User: username, ClientID: clientid, ControllerName: "UserController"}
+	log := logger.Log{ModuleName: logger.API, User: username, ClientID: clientid, ControllerName: "UserController.getUserImage"}
 	startTime := time.Now()
 	defer func() {
 		elapsed := time.Since(startTime)
 		log.PerformanceWithDuration("controllers.user.getUserImage", elapsed)
 	}()
-
-	defer func() {
-		if err := recover(); err != nil {
-			log.Error(fmt.Sprintf("getUserImage defer error: %s", err))
-			//ctx.JSON(http.StatusBadRequest, gin.H{"error": err})
-		}
-	}()
-
+	/*
+		defer func() {
+			if err := recover(); err != nil {
+				log.Error(fmt.Sprintf("getUserImage defer error: %s", err))
+				//ctx.JSON(http.StatusBadRequest, gin.H{"error": err})
+			}
+		}()
+	*/
 	log.Debug("Get User Image execution function is called.")
 
 	querystr := fmt.Sprintf(GetUserImageQuery, username)
@@ -305,6 +333,10 @@ func getUserImage(username string, clientid string) (string, error) {
 	}
 	return "", nil
 }
+
+// getUserMenus retrieves the menus for a user based on the provided parameters.
+// It takes the userID, isMobile flag, parentID, username, and clientid as input.
+// The function returns a slice of maps representing the menus and an error if any.
 
 func getUserMenus(userID int, isMobile bool, parentID int, username string, clientid string) ([]map[string]interface{}, error) {
 	log := logger.Log{ModuleName: logger.API, User: username, ClientID: clientid, ControllerName: "UserController"}
@@ -358,6 +390,15 @@ func getUserMenus(userID int, isMobile bool, parentID int, username string, clie
 
 	return jdata, nil
 }
+
+// execChangePassword is a function that handles the execution of changing a user's password.
+// It takes the following parameters:
+// - ctx: The gin.Context object for handling HTTP requests and responses.
+// - username: The username of the user whose password is being changed.
+// - oldpassword: The old password of the user.
+// - newpassword: The new password to be set for the user.
+// - clientid: The client ID associated with the user.
+// It returns an error if any error occurs during the password change process.
 
 func execChangePassword(ctx *gin.Context, username string, oldpassword string, newpassword string, clientid string) error {
 	log := logger.Log{ModuleName: logger.API, User: username, ClientID: clientid, ControllerName: "UserController"}
@@ -438,6 +479,10 @@ func execChangePassword(ctx *gin.Context, username string, oldpassword string, n
 	return nil
 }
 
+// execLogout is a function that handles the logout process for a user.
+// It takes a gin.Context object and a token string as parameters.
+// It returns a string "OK" and an error if any.
+
 func execLogout(ctx *gin.Context, token string) (string, error) {
 	log := logger.Log{ModuleName: logger.API, User: "System", ControllerName: "UserController"}
 
@@ -453,7 +498,7 @@ func execLogout(ctx *gin.Context, token string) (string, error) {
 			ctx.JSON(http.StatusBadRequest, gin.H{"error": err})
 		}
 	}()
-	_, user, clientid, err := common.GetRequestUser(ctx)
+	ID, user, clientid, err := common.GetRequestUser(ctx)
 	if err != nil {
 		log.Error(fmt.Sprintf("Get user information Error: %v", err))
 
@@ -461,10 +506,43 @@ func execLogout(ctx *gin.Context, token string) (string, error) {
 	}
 	log.ClientID = clientid
 	log.User = user
+	log.Debug(fmt.Sprintf("execLogout execution function is called. token: %s, %s ", token, ID))
+
+	Columns := []string{"LastSignOffDate", "UpdatedOn", "UpdatedBy"}
+	Values := []string{time.Now().UTC().Format("2006-01-02 15:04:05"), time.Now().UTC().Format("2006-01-02 15:04:05"), user}
+	datatypes := []int{0, 0, 0}
+	Wherestr := fmt.Sprintf("LoginName= '%s'", user)
+
+	iDBTx, err := dbconn.DB.Begin()
+	defer iDBTx.Rollback()
+
+	if err != nil {
+		log.Error(fmt.Sprintf("Begin error:%s", err.Error()))
+		ctx.JSON(http.StatusInternalServerError, "Login failed")
+	}
+
+	dboperation := dbconn.NewDBOperation(user, iDBTx, "User Logout")
+
+	index, errr := dboperation.TableUpdate(TableName, Columns, Values, datatypes, Wherestr)
+
+	if errr != nil {
+		log.Error(fmt.Sprintf("TableUpdate error:%s", errr.Error()))
+		ctx.JSON(http.StatusInternalServerError, "Logout failed")
+		return "", errr
+	}
+	log.Debug(fmt.Sprintf("index:%d", index))
+
 	config.SessionCache.Delete(ctx, token)
+
+	iDBTx.Commit()
 
 	return "OK", nil
 }
+
+// hashPassword takes a password string and returns the hashed password string.
+// It uses bcrypt.GenerateFromPassword to generate a secure hash of the password.
+// The bcrypt.DefaultCost is used to determine the cost factor of the hashing algorithm.
+// Returns the hashed password string and any error encountered during the hashing process.
 
 func hashPassword(password string) (string, error) {
 	hashedPasswordBytes, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
@@ -475,6 +553,11 @@ func hashPassword(password string) (string, error) {
 	hashedPassword := string(hashedPasswordBytes)
 	return hashedPassword, nil
 }
+
+// validatePassword validates the password for a given username.
+// It returns a boolean indicating whether the password is valid,
+// a slice of maps containing user data if the password is valid,
+// and an error if any error occurs during the validation process.
 
 func validatePassword(username string, password string) (bool, []map[string]interface{}, error) {
 	log := logger.Log{ModuleName: logger.API, User: "System", ControllerName: "UserController"}
